pyproject.toml
[project]
name = "data-clustering-app"
version = "0.1.0"
description = "Lokale Data-Clustering-App für ungelabelte Datensätze"
authors = [
    {name = "Data Clustering Team", email = "team@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "umap-learn>=0.5.4",
    "hdbscan>=0.8.33",
    "plotly>=5.15.0",
    "pyyaml>=6.0",
    "joblib>=1.3.0",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
clustering-cli = "app.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]


